requirements.txt
tqdm
pandas
pillow
tensorflow_datasets
matplotlib